+++ src/main/scala/au/csiro/variantspark/algo/package.scala
@@ -6,21 +6,24 @@ import au.csiro.variantspark.utils.CanSize
 import au.csiro.variantspark.data.Feature
 
 package object algo {
-  
-  implicit val canSplitVector = new CanSplitVector
-  implicit val canSplitArrayOfBytes = new CanSplitArrayByte
-  
-  type WideDecisionTree = DecisionTree
-  type WideDecisionTreeModel = DecisionTreeModel
-  
-  type WideRandomForest = RandomForest
-  type WideRandomForestModel = RandomForestModel
-  
-  type ByteRandomForestModel = RandomForestModel
-  type ByteRandomForest = RandomForest
+  type IndexedFeature = (Feature,Long)
    
   implicit case object CanSizeFeature extends CanSize[Feature] {
      def size(f:Feature):Int = f.size
+  }
+  
+  implicit case object CanSplitVector extends CanSplit[Vector] {
+    override def size(v:Vector) =  v.size
+    override def split(v:Vector, splitter: ClassificationSplitter, indices:Array[Int]):SplitInfo = splitter.findSplit(v.toArray, indices)
+    override def at(v:Vector)(i:Int):Double =  v(i).toDouble
+    def runtimeClass: Class[_] = classOf[Vector]
+  }
+
+  implicit case object CanSplitArrayByte extends CanSplit[Array[Byte]] {
+    override def size(v:Array[Byte]) =  v.size
+    override def split(v:Array[Byte], splitter: ClassificationSplitter, indices:Array[Int]):SplitInfo = splitter.findSplit(v, indices)
+    override def at(v:Array[Byte])(i:Int):Double =  v(i).toDouble  
+    def runtimeClass: Class[_] = classOf[Array[Byte]]
   }
   
   implicit case object CanSplitSplittable extends CanSplit[Splittable] {
@@ -29,9 +32,6 @@ package object algo {
     def split(v: Splittable, splitter: ClassificationSplitter, indices: Array[Int]): SplitInfo = v.split(splitter, indices)
     def runtimeClass: Class[_] = classOf[Splittable]
   }
-  
-  
-  
+    
   implicit def toTreeFeatueRDD(rdd:RDD[TreeFeature]) = new TreeFeatueRDDFunction(rdd)
-
 }
